\name{DeducerRDSAnalyst-internal}
\title{Internal DeducerRDSAnalyst Objects}
% terms
\alias{JavaGDL}

%
% These need to be documented!
%

\description{
  Internal DeducerRDSAnalyst functions.
}
%\usage{
%anova.DeducerRDSAnalyst(object, \dots)
%}
\details{
  Most of these are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :).
}
%\arguments{
% \item{\dots}{further arguments passed to or used by methods.}
%}
\seealso{Deducer}
\keyword{internal}
