% Generated by roxygen2 (4.0.1): do not edit by hand
\name{save_str_trim}
\alias{save_str_trim}
\title{Safely Trim whitespace from start and end of string.}
\usage{
save_str_trim(x)
}
\arguments{
\item{x}{input character vector}
}
\value{
character vector with leading and trailing whitespace removed
}
\description{
Safely Trim whitespace from start and end of string.
}
\details{
This is a version of str_tirm from the stringr package that returns its argument when the argument is not a string.
}
\seealso{
str_trim
}

