% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wtd.contingency.tables}
\alias{wtd.contingency.tables}
\title{Weighted Contingency Tables}
\usage{
wtd.contingency.tables(row.vars, col.vars, stratum.var, weights = NULL,
  data = NULL, missing.include = FALSE, subset = NULL)
}
\arguments{
\item{row.vars}{Row variables.}

\item{col.vars}{Column variables.}

\item{stratum.var}{A varaible to stratify by.}

\item{weights}{Sampling weights.}

\item{data}{A data.frame.}

\item{missing.include}{Include missing values as a level.}

\item{subset}{A subset expression.}
}
\description{
Weighted Contingency Tables
}

