% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wtd.descriptive.table}
\alias{wtd.descriptive.table}
\title{Weighted descriptives}
\usage{
wtd.descriptive.table(vars, strata, data, func.names = c("Mean",
  "St. Deviation", "Median", "25th Percentile", "75th Percentile", "Minimum",
  "Maximum", "Population Size"), weights, subset = NULL)
}
\arguments{
\item{vars}{The variables}

\item{strata}{The stratifying variables}

\item{data}{A data.frame}

\item{func.names}{The descriptive statistics to calculate}

\item{weights}{The weights}

\item{subset}{A logical expression for subsetting}
}
\description{
Weighted descriptives
}

