% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wtd.frequencies}
\alias{wtd.frequencies}
\title{Weighted frequencies}
\usage{
wtd.frequencies(data, weights = NULL, r.digits = 1)
}
\arguments{
\item{data}{A data.frame of discrete variables.}

\item{weights}{The weights}

\item{r.digits}{The number of significant digits to display}
}
\description{
Weighted frequencies
}

